Task-9_Data Migration/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-cosmos
pyodbc
azure-functions
azure-identity
